modified general layout
