--- src/assets/scss/Header.style.scss
@@ -1,58 +1,0 @@
-/* Effect 1: Brackets */
-.link-spacing {
-  margin: "0px 20px";
-}
-
-.cl-effect-1 a:hover {
-  text-decoration: none;
-}
-
-.cl-effect-1 a::before,
-.cl-effect-1 a::after {
-  display: inline-block;
-  opacity: 0;
-  -webkit-transition: -webkit-transform 0.3s, opacity 0.2s;
-  -moz-transition: -moz-transform 0.3s, opacity 0.2s;
-  transition: transform 0.3s, opacity 0.2s;
-}
-
-.cl-effect-1 a::before {
-  margin-right: 10px;
-  content: "{";
-  -webkit-transform: translateX(20px);
-  -moz-transform: translateX(20px);
-  transform: translateX(20px);
-}
-
-.cl-effect-1 a::after {
-  margin-left: 10px;
-  content: "}";
-  -webkit-transform: translateX(-20px);
-  -moz-transform: translateX(-20px);
-  transform: translateX(-20px);
-}
-
-.cl-effect-1 a:hover::before,
-.cl-effect-1 a:hover::after,
-.cl-effect-1 a:focus::before,
-.cl-effect-1 a:focus::after {
-  opacity: 1;
-  -webkit-transform: translateX(0px);
-  -moz-transform: translateX(0px);
-  transform: translateX(0px);
-}
-
-/* Active Link Effect */
-.active a:hover {
-  text-decoration: none;
-}
-
-.active a::before {
-  margin-right: 10px;
-  content: "[";
-}
-
-.active a::after {
-  margin-left: 10px;
-  content: "]";
-}
