--- src/components/elements/BlogCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './BlogCard';
