--- src/components/elements/CommentCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './CommentCard';
