--- src/components/elements/Logo/index.js
@@ -1,1 +1,0 @@
-export { default } from './Logo.js';
