--- src/components/elements/Rose/index.js
@@ -1,1 +1,0 @@
-export { default } from './Rose.js';
