+++ src/views/Home.js
@@ -1,16 +1,12 @@
 import React from "react";
 
-import Logo from "../components/elements/Logo";
+import Blog from "../components/sections/Blog";
 
 export default function Home() {
   return (
-    <>
-      <h1>Hello, I&apos;m Edward Nunez</h1>
-      <p>
-        Passionate and experienced Full Stack Software Engineer, with cloud
-        solutions and automation background.
-      </p>
-      <Logo />
-    </>
+    <div className="row">
+      <Blog />
+      <div className="col-md-3 ml-auto">side bar</div>
+    </div>
   );
 }
