+++ src/views/Work.js
@@ -2,8 +2,8 @@ import React from "react";
 
 export default function Work() {
   return (
-    <>
+    <div className="row">
       <h1>Work</h1>
-    </>
+    </div>
   );
 }
